Remove mocker lib
